fix(dashboard): Elimina type casting de frontend para prevenir crashes de hidratacion
ANÁLISIS DE ERROR FINAL:
- Se presentaba un TypeError durante el build frontend (Reflex Export) al evaluar la condición 'dias <= 30' debido a que el elemento iterado de 'rx.foreach' no infería implícitamente el tipo de dato subyacente de la clave del diccionario ('dias_restantes'), tratándolo como genérico ('ObjectItemOperation') en vez de 'int'.
- Corrección implementada: Se agregó explícitamente el casting a entero mediante el método '.to(int)' en el diccionario antes de pasarlo al componente 'badge_vencimiento', permitiendo así la evaluación polinómica sin fallos.
- Se restableció la importación omitida por error ('top_asesores_chart') en 'dashboard.py'.

+++ src/presentacion_reflex/components/dashboard/vencimientos_tables.py
@@ -52,7 +52,7 @@ def _tabla_vencimientos(titulo: str, icon: str, color_scheme: str, lista_estado)
                                     rx.table.cell(rx.text(item["parte_contratante"], size="2", weight="medium")),
                                     rx.table.cell(rx.text(item["direccion"], size="2", color=styles.TEXT_SECONDARY)),
                                     rx.table.cell(rx.text(item["fecha_fin"], size="2")),
-                                    rx.table.cell(_vencimiento_badge(item["dias_restantes"])),
+                                    rx.table.cell(_vencimiento_badge(item["dias_restantes"].to(int))),
                                     align="center",
                                 )
                             )
